+++ R/wrapr_exports.R
@@ -1,7 +1,20 @@
 
+#' @importFrom wrapr apply_left
+#' @export
+wrapr::apply_left
+
+#' @importFrom wrapr apply_right
+#' @export
+wrapr::apply_right
+
+
 #' @importFrom wrapr mk_tmp_name_source
 #' @export
 wrapr::mk_tmp_name_source
+
+#' @importFrom wrapr map_to_char
+#' @export
+wrapr::map_to_char
 
 #' @importFrom wrapr %.>%
 #' @export
@@ -15,6 +28,14 @@ wrapr::let
 #' @export
 wrapr::qc
 
+#' @importFrom wrapr qe
+#' @export
+wrapr::qe
+
+#' @importFrom wrapr qae
+#' @export
+wrapr::qae
+
 #' @importFrom wrapr build_frame
 #' @export
 wrapr::build_frame
@@ -26,3 +47,8 @@ wrapr::draw_frame
 #' @importFrom wrapr qchar_frame
 #' @export
 wrapr::qchar_frame
+
+#' @importFrom wrapr %:=%
+#' @export
+wrapr::`%:=%`
+